refactored web-literacy.jsx
